+++ sample-app/src/pages/UniversalSearchPage.tsx
@@ -3,7 +3,7 @@ import { useLayoutEffect } from 'react';
 import { useAnswersActions } from '@yext/answers-headless-react';
 import '../sass/UniversalSearchPage.scss';
 import { UniversalResultsConfig } from '../universalResultsConfig';
-import SearchHandler from '../utils/searchhandler';
+import { executeSearchWithIntents, getSearchIntents } from '../utils/search-operations';
 
 const universalResultsFilterConfig = {
   show: true
@@ -19,8 +19,8 @@ export default function UniversalSearchPage(props: { universalResultsConfig: Uni
     })
     answersActions.setVerticalKey('');
     const executeQuery = async () => {
-      const searchIntents = await SearchHandler.getSearchIntents(answersActions, false);
-      SearchHandler.executeSearchWithIntents(answersActions, false, searchIntents || []);
+      const searchIntents = await getSearchIntents(answersActions, false);
+      executeSearchWithIntents(answersActions, false, searchIntents || []);
     };
     executeQuery();
   }, [answersActions]);
